refactor: restructure charts module
