+++ packages/charts/src/api/types.ts
@@ -83,8 +83,3 @@ export type Chart = {
 export interface ChartsIndexResponse {
   charts: Chart[];
 }
-
-export interface NavigraphCharts {
-  getChartsIndex: typeof import("./lib/getChartsIndex").default;
-  getChartImage: typeof import("./lib/getChartImage").default;
-}
