+++ packages/charts/src/index.ts
@@ -1,3 +1,4 @@
-export * from "./api";
-export * from "./public-types";
-export * from "./lib";
+export * from "./api/types";
+export * from "./util";
+
+export { getChartsAPI } from "./lib/getChartsAPI";
