+++ packages/charts/src/lib/getChartsAPI.ts
@@ -1,10 +1,9 @@
 import { getApp, Logger, NotInitializedError, Scope } from "@navigraph/app";
-import getChartImage from "./lib/getChartImage";
-import getChartsIndex from "./lib/getChartsIndex";
-import { NavigraphCharts as NavigraphChartsAPI } from "./public-types";
+import getChartImage from "../api/getChartImage";
+import getChartsIndex from "../api/getChartsIndex";
 
 /** Grabs a reference to an object containing available Navigraph Charts API functionality */
-export const getChartsAPI = (): NavigraphChartsAPI => {
+export const getChartsAPI = () => {
   const app = getApp();
 
   if (!app) {
